refactor(gateway): extract @plot/gateway from @plot/cli
The operator-console HTTP server (Bun.serve routes, SSE streams, projection
replay, transcript serving, embedded SPA assets) gets its own package above
registry/projection/session. The CLI keeps thin web/api commands and passes
process concerns in via existing options; it no longer depends on
@plot/projection. scripts/web-assets.ts now generates the asset module into
packages/gateway. Verified with a local release build: the Bun single-file
binary still embeds the SPA through the workspace dependency.
