+++ packages/gateway/package.json
@@ -1,13 +1,13 @@
 {
-	"name": "@plot/cli",
+	"name": "@plot/gateway",
 	"private": true,
-	"bin": {
-		"plot": "./src/main.ts"
-	},
 	"type": "module",
+	"sideEffects": false,
 	"exports": {
-		"./cli": "./src/cli.ts",
-		"./runtime": "./src/runtime.ts"
+		".": {
+			"types": "./src/gateway.ts",
+			"import": "./src/gateway.ts"
+		}
 	},
 	"scripts": {
 		"typecheck": "tsc -p tsconfig.json --noEmit",
@@ -16,11 +16,9 @@
 	},
 	"dependencies": {
 		"@plot/common": "workspace:*",
-		"@plot/gateway": "workspace:*",
+		"@plot/projection": "workspace:*",
 		"@plot/registry": "workspace:*",
-		"@plot/session": "workspace:*",
-		"@plot/tui": "workspace:*",
-		"citty": "^0.1.6"
+		"@plot/session": "workspace:*"
 	},
 	"devDependencies": {
 		"@repo/tsconfig": "workspace:*"
