+++ packages/gateway/src/web-assets.generated.ts
